# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GURU - Agent Testing Platform API",
    "description": "API for managing AI agents, projects, iterations, teams, organizations, and billing.",
    "version": "1.0.0",
    "contact": {
      "name": "GURU Support"
    }
  },
  "servers": [
    {
      "url": "https://guru-api-production.up.railway.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/settings/profile": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get current user profile",
        "responses": {
          "200": { "description": "User profile data" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update user profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_name": { "type": "string" },
                  "company": { "type": "string" },
                  "job_title": { "type": "string" },
                  "avatar_url": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated profile" }
        }
      }
    },
    "/api/settings/change-password": {
      "post": {
        "tags": ["Settings"],
        "summary": "Change password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["current_password", "new_password"],
                "properties": {
                  "current_password": { "type": "string" },
                  "new_password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password changed" },
          "401": { "description": "Current password incorrect" }
        }
      }
    },
...
    "/api/settings/verify-email": {
      "post": {
        "tags": ["Settings"],
        "summary": "Verify email with token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": {
                  "token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Email verified" },
          "404": { "description": "Invalid token" },
          "410": { "description": "Token expired" }
        }
      }
    },
...
    "/api/settings/account": {
      "delete": {
        "tags": ["Settings"],
        "summary": "Delete user account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account deleted" },
          "400": { "description": "Must transfer org ownership first" }
        }
      }
    },
...
  }
}